refactor: event page
